requirements.txt
requests~=2.32.3
fastapi
uvicorn
httpx
python-dotenv